rerfactor: attempt to decrease optin action response time
